regulai/requirements.txt
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
Django==4.2.28
django-cors-headers==4.9.0
djangorestframework==3.16.1
docstring_parser==0.17.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
lxml==6.0.2
openai==1.102.0
openpyxl==3.1.5
packaging==26.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
yara-python==4.5.4
